+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # Changelog
 
-### 1.0 (2024.11.07)
+### 1.0.0 (2024.11.07)
 - Initial release.
+
+### 1.0.1 (2025.01.24)
+- Added new event `add_shipping_info`
+- Added plugin settings page
+- Added user-friendly events log
+- Added user-friendly error log